Separate `__new__` and `__init__` (#44)
Previously, I had written both allocation and initialisation code in `__new__`, because I thought I would have to `Py_DECREF` the object before exiting if I encountered an error while parsing the arguments. Having allocation and initialisation in the same function allowed me to parse the arguments first, and then initialise, thereby letting me skip allocation and deallocation in case of errors.

However, it is not necessary to `Py_DECREF` the object if an error is encountered in `__init__`, because returning an error code causes the newly-created object to not get bound to a name in Python land, thereby decrementing its reference count back to 0, and running the deallocator. When I found this out, I decided to implement this pattern (i.e. allocation in `__new__` and initialisation in `__init__`), because the CPython documentation recommends it.