Use newer Setuptools on Windows to allow developing on MSYS2 (#85)
MSYS2 provides a patched version of Python, which has a non-standard platform tag. The version of Setuptools which plays nice with the Apple Linker when cibuildwheel builds PyPy wheels is 70.0.0. This version rejects the non-standard platform tag, but version 71.0.0 allows it. Hence, use the latter version when building on Windows.

This change purely aids development on MSYS2. MSYS2 wheels won't be uploaded to PyPI. (They will actually be rejected because of the platform tag.)