Custom fonts in project website (#145)
Signika has absolutely phenomenal hinting and anti-aliasing, and I have been using it as the UI font on Linux and Windows for a while now. It is objectively superior to Arial, which has thin strokes, which bleed into the background colour. It is a good choice for the website. JetBrains Mono is likewise better than Liberation Mono.

I made the decision to self-host the fonts (using Git LFS) because loading them lead to the text bouncing around as the browser used the fallback font, and then switched to the first preference once it was downloaded. However, since I am hosting only a subset of the font files, I must not use the same font names to comply with the SIL Open Font Licence version 1.1. Hence, I stuck 'pysorteddict' to the original names.