Replace Python method names with C++
It is not possible to benchmark the insertion or deletion of a single value becuse those are mutating operations. Hence, I'll benchmark insertion and immediate deletion, which together are non-mutating. (Any changes occurring under the hood are okay; they will simulate the behaviour of the sorted dictionary under heavy load.)

In Python, `__setitem__` is called for insertion, and `__delitem__`, for deletion. It is thus wrong to use '__setitem__' when benchmarking insertion and immediate deletion. In C++, both operations are performed by `setitem`, so using 'setitem' in the Jupyter notebook and the names of the files is correct. In anticipation of these benchmarks, which I am yet to add, I am making the related changes for the existing benchmarks.