Add Visual Studio Code configuration file
I want to keep this file because I dislike seeing error squiggles under Python functions. The autocompletion is also nice. I was hesitant to commit it because it includes some OS-specific details, which means I would have to change it if I wanted to work on another system. However, I have already committed a Jupyter notebook which also contains some such details, so might as well do so for this file, too.