Use pi symbol in homepage (#206)
Downloaded the files for the bold and regular styles of Signika with all available character sets. Picked out the the glyphs for pi from them and put them into the files checked into this repository using FontForge. Without this, a fallback font was used to render it. Didn't like that.