Specify Python version when building website
For reasons not completely clear to me, this is required for the Pyodide kernel to be available in the GitHub Pages deployment. I have tested this in a separate repository, and am hoping that it works correctly here. I have also added the workflow file to the sensitivity list so that the workflow gets triggered when it is changed.